Move node and networking related helpers from pkg/util to component helpers
