+++ pkg/kubelet/dockershim/network/testing/plugins_test.go
@@ -25,10 +25,10 @@ import (
 	"testing"
 
 	utilsets "k8s.io/apimachinery/pkg/util/sets"
+	sysctltest "k8s.io/component-helpers/node/utils/sysctl/testing"
 	kubeletconfig "k8s.io/kubernetes/pkg/kubelet/apis/config"
 	kubecontainer "k8s.io/kubernetes/pkg/kubelet/container"
 	"k8s.io/kubernetes/pkg/kubelet/dockershim/network"
-	sysctltest "k8s.io/kubernetes/pkg/util/sysctl/testing"
 	netutils "k8s.io/utils/net"
 
 	"github.com/golang/mock/gomock"
